[#6]Rename(ranking): 통일성 위해 domain -> entity 패키지 이름 수
