+++ src/main/java/today/sesac/shoutify/ranking/entity/Ranking.java
@@ -1,4 +1,4 @@
-package today.sesac.shoutify.ranking.domain;
+package today.sesac.shoutify.ranking.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
