+++ src/main/java/today/sesac/shoutify/ranking/repository/RankingRepository.java
@@ -3,7 +3,7 @@ package today.sesac.shoutify.ranking.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import today.sesac.shoutify.ranking.domain.Ranking;
+import today.sesac.shoutify.ranking.entity.Ranking;
 
 /**
  * 순위(랭킹) 정보를 관리하는 레포지토리입니다.
